% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{add_facet}
\alias{add_facet}
\title{Facet a plot according to specified variables}
\usage{
add_facet(p, facet_terms = NULL, facet_scales = "fixed",
  facet_space = "fixed")
}
\arguments{
\item{p}{[ggplot object] A plot object we might want to facet.}

\item{facet_terms}{[list of strings] The names of the columns of the data
used in generating p which we want to facet by. Defaults to NULL, in which
case no faceting is applied.}
}
\value{
p [ggplot object] The version of p including potential faceting.
}
\description{
Facet a plot according to specified variables
}

