% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{ggboxplot}
\alias{ggboxplot}
\title{Nice default options for (grid of) boxplots}
\usage{
ggboxplot(plot_data, opts = list())
}
\arguments{
\item{plot_data}{[data.frame] A data.frame with data to create the boxplot.
The variables for x, y, colors / fill types, and faceting must be provided
by opts, see merge_boxplot_opts().}

\item{opts}{[list] A partially specified list used to customize the
appareance in ggplot theme(). Options that are already specified will not
be changed, those that are not will be filled in with defaults.}
}
\value{
p [ggplot] A ggplot geom_boxplot() with nice defaults
}
\description{
Nice default options for (grid of) boxplots
}
\examples{
# create an example data set
X <- data.frame(
 "x" = sample(c("A", "B", "C"), 1000, replace = TRUE),
 "y" = rnorm(1000),
 "fill" = sample(c("alpha", "beta"), 1000, replace = TRUE)
)

ggboxplot(X)
ggboxplot(X, list(col = "fill", fill = "fill"))
}

