% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{ggcontours}
\alias{ggcontours}
\title{Default Contour / Stat Density 2D plot}
\usage{
ggcontours(plot_data, opts = list())
}
\arguments{
\item{plot_data}{[data.frame] A data.frame with data to create the density plot
from. The variables for x, y axis, fill, and faceting must be
provided by opts. Otherwise, they will be filled in with defaults, see
merge_contour_opts(). The order in which variables appear on the axes is
controlled by opts$x_order and opts$y_order.}

\item{opts}{[list] A partially specified list used to customize appearance in
ggplot theme(). Options that are already specified will not be changed,
those that are not will be filled in with defaults.}
}
\value{
p [ggplot] A ggplot stat_density2d object with nice defaults.
}
\description{
This is a default for visualizing 2d density plots.
}
\examples{
plot_data <- expand.grid(
  "ix" = seq_len(1000),
  "row" = c("A", "B", "C"),
  "col" = c("alpha", "beta", "gamma")
)

plot_data <- cbind(
  plot_data,
  x = rnorm(9000),
  y = rnorm(9000)
)
ggcontours(plot_data, list(facet_terms = c("row", "col")))
ggcontours(plot_data, list(facet_terms = c("row", ".")))
ggcontours(plot_data, list(facet_terms = c("row", "."), fill_type = "discrete", fill = "col"))
}

