% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Make heatmaps with nice defaults}
\usage{
ggheatmap(plot_data, opts = list())
}
\arguments{
\item{plot_data}{[data.frame] A data.frame with data to create the heatmap
from. The variables for x, y axis, filling, and faceting must be provided
by opts. Otherwise, they will be filled in with defaults, see
merge_heatmap_opts(). The order in which variables appear on the axes is
controlled by opts$x_order and opts$y_order.}

\item{opts}{[list] A partially specified list used to customize appearance in
ggplot theme(). Options that are already specified will not be changed,
those that are not will be filled in with defaults.}
}
\value{
p [ggplot] A ggplot geom_tile() heatmap with nice defaults.
}
\description{
Make heatmaps with nice defaults
}
\examples{
X <- matrix(rnorm(10 * 10), 10, 10, dimnames = list(1:10, letters[1:10]))
plot_data <- melt(X, varnames = c("x", "y"), value.name = "fill")
opts <- list(y_order = letters[10:1])
ggheatmap(plot_data, opts)
}

