% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{gglines}
\alias{gglines}
\title{Make line plots with nice defaults}
\usage{
gglines(plot_data, opts = list())
}
\arguments{
\item{plot_data}{[data.frame] A data.frame with data to create the lines plot
from. The variables for x, y axis, colors / linetype, and faceting must be
provided by opts. Otherwise, they will be filled in with defaults, see
merge_line_opts(). The order in which variables appear on the axes is
controlled by opts$x_order and opts$y_order.}

\item{opts}{[list] A partially specified list used to customize appearance in
ggplot theme(). Options that are already specified will not be changed,
those that are not will be filled in with defaults.}
}
\value{
p [ggplot] A ggplot geom_tile() heatmap with nice defaults.
}
\description{
Make line plots with nice defaults
}
\examples{
data(lines)
opts <- list(col = "color", facet_terms = c("rows", "columns"),
             color_scale = scale_color_brewer(palette = "Set3"))
gglines(plot_data, opts)
}

