% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{merge_theme_opts}
\alias{merge_theme_opts}
\title{Merge in Theme Defaults}
\usage{
merge_theme_opts(opts = list())
}
\arguments{
\item{opts}{[list] A partially specified list used to customize themes in
ggplot theme(). Options that are already specified will not be changed,
those that are not will be filled in with defaults.}
}
\value{
opts [list]  A version of opts with unspecified options filled in
  with defaults.
}
\description{
For a partially specified list opts, this merges in defaults
for those options that are not specified.
}

