% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{min_theme}
\alias{min_theme}
\title{Provide a Custom, Minimal Theme}
\usage{
min_theme(opts = list())
}
\arguments{
\item{opts}{[list] A partially specified list used to customize themes in
ggplot theme(). Options that are already specified will not be changed,}
}
\value{
theme_obj [ggplot theme] A theme object that can be used to modify
  the appearance of a ggplot object.
}
\description{
Provide a Custom, Minimal Theme
}

