% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{order_vars}
\alias{order_vars}
\title{Reorder levels in a vector}
\usage{
order_vars(x, x_order = NULL)
}
\arguments{
\item{x}{[vector] A vector whose factors we want to reorder (or create).}

\item{x_order}{[vectors] The new levels to use for each name in vars.}
}
\value{
x [vector] version of x with levels in the correct order
}
\description{
Reorder levels in a vector
}

