% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{order_vars}
\alias{order_vars}
\title{Reorder A collection of factors in a data.frame}
\usage{
order_vars(X, vars, var_orders)
}
\arguments{
\item{X}{[data.frame] A data.frame whose columns have levels we want to
reorder.}

\item{vars}{[numeric vector] A vector of column names in X, for each column
that we want to reorder.}

\item{vars}{[list of vectors] The new levels to use for each name in vars.
If this has any nulls, we use the existing levels, or just the unique
values.}
}
\value{
X [data.frame] A version of X with the columns in vars already
  reordered.
}
\description{
Reorder A collection of factors in a data.frame
}

